% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{Birthday}
\alias{Birthday}
\title{Title birthday function}
\usage{
Birthday(x)
}
\arguments{
\item{x}{x as a vector}
}
\value{
returns
}
\description{
Title birthday function
}
\examples{
Birthday(20:24)

}
