% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Dataset}
\format{
A data frame with however-many rows and columns:
\describe{
\item{DDT}{numeric: DDT concentration}
\item{...}{(add other variable descriptions if needed)}
}
}
\usage{
data(ddt)
}
\description{
A dataset containing DDT concentration measurements.
}
\keyword{dataset}
