% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damages Data (FIREDAM.csv)}
\format{
\subsection{\code{fire}}{

A data frame with X rows and Y columns:
\describe{
\item{Column1}{Short description of column 1}
\item{Column2}{Short description of column 2}
\item{Column3}{Short description of column 3}
\item{Column4}{Short description of column 4}
\% Add one item per column based on names(fire)
}
}
}
\source{
Course dataset: FIREDAM.csv (provided for ILAS2019 labs)
}
\usage{
fire
}
\description{
Fire Damages Data (FIREDAM.csv)
}
\keyword{datasets}
