% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of simulated experiments. Default is 100.}

\item{n}{Number of trials per experiment. Default is 10.}

\item{p}{Probability of success for each trial. Default is 0.5.}
}
\value{
A named numeric vector giving the proportion of each possible
}
\description{
Title
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
mybin(iter = 1000, n = 18, p = 0.3)
}
