% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFIrst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My First Function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A Numeric Vector.}
}
\value{
A list with component x and y where y is the square of 'x'.
}
\description{
My First Function
}
\examples{
myfirstfunction(1:10)
}
